Fix bug in montecarlo_common.f90:flip_h subroutine (heatbath)
Fix a bug in calculating the polar coordinates of a spin in
the flip_h subroutine, used in heatbath calculations.

This bug caused spins which were close to the -1 z-pole
to incorrectly flip to +1 z-pole.
This happened in the if clause where zstheta < 1.d-2. The if clause
incorrectly assumed that zctheta is close to +1 in that case, while
it can also be close to -1.